requirements.txt
Django>=4.2,<5.1
djangorestframework
djangorestframework-simplejwt
python-dotenv
psycopg2-binary
pytest
pytest-django
requests
hiero-sdk-python
pythclient
celery
redis